% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpreg_functions.R
\name{betaRound}
\alias{betaRound}
\title{betaRound}
\usage{
betaRound(a)
}
\arguments{
\item{a}{A vector}
}
\value{
rounded vector
}
\description{
Rounds each element in the vector (to 0 or 1) based off of a threshold (abs value <0.2)
}
\examples{
betaRound(c(1.30,0.02,-1.17))

}
