% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpreg_functions.R
\name{L2diff}
\alias{L2diff}
\title{L2diff}
\usage{
L2diff(a, b)
}
\arguments{
\item{a}{A vector}

\item{b}{A vector}
}
\value{
The l2 difference of the vectors
}
\description{
Calculates the l2 difference between two vectors of the same length
}
\examples{
L2diff(c(1,2,3,4),c(3,6,9,3))

}
