% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpreg_functions.R
\name{SignMatch}
\alias{SignMatch}
\title{Signmatch}
\usage{
SignMatch(a, b)
}
\arguments{
\item{a}{A vector}

\item{b}{A vector}
}
\value{
A matrix of True/False positive matches.
}
\description{
Takes pairs of elements between two vectors and sees whether or not both elements are in the same category (abs value less than 0.2, abs value greater than 0.2).
Calculates percentage of pairs that are in the same category
}
\examples{
SignMatch(c(1,2,3,4),c(3,6,9,3))

}
