% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpreg_functions.R
\name{thresh_within}
\alias{thresh_within}
\title{thresh_within}
\usage{
thresh_within(grp.ind, est, delta)
}
\arguments{
\item{grp.ind}{a p-dim vector of group indices, e.g. 1, 1, 1, 2, 2, 3, 3, 3, 4, 4}

\item{est}{a p-dim vector (beta-hat) containing the group lasso solution}

\item{delta}{scalar, controlling the amount of threshold}
}
\value{
est.out
}
\description{
Runs thresholded group lasso on group lasso output. Created by: Sumanta Basu
}
\examples{
thresh_within(grp.ind = c(0,0,1,1,2,2,3), est =as.matrix(c(1,2,3,4,7,6,5),ncol=1),delta = 0.01)

}
